week1-WatchThis/requirements.txt
pandas
numpy
requests
sentence_transformers
psycopg2-binary 
pgvector